// projects/10/JackAnalyzer/types/Keyword.ts
enum Keyword {
    Class = "class",
    Method = "method",
    Function = "function",
    Constructor = "constructor",
    Int = "int",
    Boolean = "boolean",
    Char = "char",
    Void = "void",
    Var = "var",
    Static = "static",
    Field = "field",
    Let = "let",
    Do = "do",
    If = "if",
    Else = "else",
    While = "while",
    Return = "return",
    True = "true",
    False = "false",
    Null = "null",
    This = "this",
}

export default Keyword;